Update ValueSecretSchema documentation
