--- pkg/resources/format_spec.go
@@ -1,154 +1,0 @@
-package resources
-
-import (
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
-)
-
-func FormatSpecSchema(elem string, description string, required bool) *schema.Schema {
-	return &schema.Schema{
-		Type: schema.TypeList,
-		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{
-				"avro": {
-					Description: "Avro format.",
-					Type:        schema.TypeList,
-					Optional:    true,
-					ForceNew:    true,
-					MaxItems:    1,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"schema_registry_connection": IdentifierSchema("schema_registry_connection", "The name of a schema registry connection.", true),
-							"key_strategy": {
-								Description:  "How Materialize will define the Avro schema reader key strategy.",
-								Type:         schema.TypeString,
-								Optional:     true,
-								ForceNew:     true,
-								ValidateFunc: validation.StringInSlice(strategy, true),
-							},
-							"value_strategy": {
-								Description:  "How Materialize will define the Avro schema reader value strategy.",
-								Type:         schema.TypeString,
-								Optional:     true,
-								ForceNew:     true,
-								ValidateFunc: validation.StringInSlice(strategy, true),
-							},
-						},
-					},
-				},
-				"protobuf": {
-					Description: "Protobuf format.",
-					Type:        schema.TypeList,
-					Optional:    true,
-					ForceNew:    true,
-					MaxItems:    1,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"schema_registry_connection": IdentifierSchema("schema_registry_connection", "The name of a schema registry connection.", true),
-							"message": {
-								Description: "The name of the Protobuf message to use for the source.",
-								Type:        schema.TypeString,
-								Required:    true,
-								ForceNew:    true,
-							},
-						},
-					},
-				},
-				"csv": {
-					Description: "CSV format.",
-					Type:        schema.TypeList,
-					Optional:    true,
-					ForceNew:    true,
-					MaxItems:    2,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"column": {
-								Description: "The columns to use for the source.",
-								Type:        schema.TypeInt,
-								Optional:    true,
-								ForceNew:    true,
-							},
-							"delimited_by": {
-								Description: "The delimiter to use for the source.",
-								Type:        schema.TypeString,
-								Optional:    true,
-								ForceNew:    true,
-							},
-							"header": {
-								Description: "The number of columns and the name of each column using the header row.",
-								Type:        schema.TypeList,
-								Elem:        &schema.Schema{Type: schema.TypeString},
-								Optional:    true,
-								ForceNew:    true,
-							},
-						},
-					},
-				},
-				"json": {
-					Description: "JSON format.",
-					Type:        schema.TypeBool,
-					Optional:    true,
-					ForceNew:    true,
-				},
-				"text": {
-					Description: "Text format.",
-					Type:        schema.TypeBool,
-					Optional:    true,
-					ForceNew:    true,
-				},
-			},
-		},
-		Required:    required,
-		Optional:    !required,
-		MinItems:    1,
-		MaxItems:    1,
-		ForceNew:    true,
-		Description: description,
-	}
-}
-
-func SinkFormatSpecSchema(elem string, description string, required bool) *schema.Schema {
-	return &schema.Schema{
-		Type: schema.TypeList,
-		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{
-				"avro": {
-					Description: "Avro format.",
-					Type:        schema.TypeList,
-					Optional:    true,
-					ForceNew:    true,
-					MaxItems:    1,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"schema_registry_connection": IdentifierSchema("schema_registry_connection", "The name of a schema registry connection.", true),
-							"avro_key_fullname": {
-								Description: "The full name of the Avro key schema.",
-								Type:        schema.TypeString,
-								Optional:    true,
-								ForceNew:    true,
-							},
-							"avro_value_fullname": {
-								Description: "The full name of the Avro value schema.",
-								Type:        schema.TypeString,
-								Optional:    true,
-								ForceNew:    true,
-							},
-						},
-					},
-				},
-				"json": {
-					Description: "JSON format.",
-					Type:        schema.TypeBool,
-					Optional:    true,
-					ForceNew:    true,
-				},
-			},
-		},
-		Required:    required,
-		Optional:    !required,
-		MinItems:    1,
-		MaxItems:    1,
-		ForceNew:    true,
-		Description: description,
-	}
-}
